requirements.txt
fastapi
uvicorn
twilio
python-dotenv
openai